agrovision-azure-func/ml-weed/requirements.txt
azure-functions
azure-cognitiveservices-vision-customvision
azure-storage-blob
azure-cosmos